readmodel: extract shared experiment and frontier projections
