+++ internal/cli/read_classification.go
@@ -5,112 +5,37 @@ import (
 	"time"
 
 	"github.com/bytter/autoresearch/internal/entity"
+	"github.com/bytter/autoresearch/internal/readmodel"
 	"github.com/bytter/autoresearch/internal/store"
 )
 
 const (
-	experimentClassificationLive = "live"
-	experimentClassificationDead = "dead"
+	experimentClassificationLive = readmodel.ExperimentClassificationLive
+	experimentClassificationDead = readmodel.ExperimentClassificationDead
 )
 
-type experimentReadClass struct {
-	Classification   string `json:"classification"`
-	HypothesisStatus string `json:"hypothesis_status,omitempty"`
-}
-
-type experimentReadView struct {
-	*entity.Experiment
-	Classification   string `json:"classification"`
-	HypothesisStatus string `json:"hypothesis_status,omitempty"`
-}
-
-type staleExperimentView struct {
-	ID            string  `json:"id"`
-	Hypothesis    string  `json:"hypothesis"`
-	Status        string  `json:"status"`
-	LastEventKind string  `json:"last_event_kind"`
-	StaleMinutes  float64 `json:"stale_minutes"`
-}
-
-func normalizeExperimentReadClass(class experimentReadClass) experimentReadClass {
-	if class.Classification == "" {
-		class.Classification = experimentClassificationLive
-	}
-	return class
-}
-
-// loopActionable reports whether the research loop should still steer from
-// this experiment's current hypothesis state. This is intentionally broader
-// than gc's "terminal" notion: decisive-but-unreviewed hypotheses are also
-// non-actionable for further loop steering even though they are not yet safe
-// to reclaim.
-func (class experimentReadClass) loopActionable() bool {
-	return normalizeExperimentReadClass(class).Classification == experimentClassificationLive
-}
+type experimentReadClass = readmodel.ExperimentReadClass
+type experimentReadView = readmodel.ExperimentReadView
+type staleExperimentView = readmodel.StaleExperimentView
 
 func experimentReadClassForID(classByID map[string]experimentReadClass, expID string) experimentReadClass {
-	return normalizeExperimentReadClass(classByID[expID])
+	return readmodel.ExperimentReadClassForID(classByID, expID)
 }
 
 func classifyExperimentsForRead(s *store.Store, exps []*entity.Experiment) (map[string]experimentReadClass, error) {
-	hyps, err := s.ListHypotheses()
-	if err != nil {
-		return nil, err
-	}
-	hypStatus := make(map[string]string, len(hyps))
-	for _, h := range hyps {
-		if h == nil {
-			continue
-		}
-		hypStatus[h.ID] = h.Status
-	}
-	out := make(map[string]experimentReadClass, len(exps))
-	for _, e := range exps {
-		if e == nil {
-			continue
-		}
-		out[e.ID] = classifyExperimentForRead(e, hypStatus)
-	}
-	return out, nil
+	return readmodel.ClassifyExperimentsForRead(s, exps)
 }
 
 func annotateExperimentsForRead(s *store.Store, exps []*entity.Experiment) ([]*experimentReadView, error) {
-	classByID, err := classifyExperimentsForRead(s, exps)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]*experimentReadView, 0, len(exps))
-	for _, e := range exps {
-		if e == nil {
-			continue
-		}
-		class := normalizeExperimentReadClass(classByID[e.ID])
-		out = append(out, &experimentReadView{
-			Experiment:       e,
-			Classification:   class.Classification,
-			HypothesisStatus: class.HypothesisStatus,
-		})
-	}
-	return out, nil
+	return readmodel.AnnotateExperimentsForRead(s, exps)
 }
 
 func annotateExperimentForRead(s *store.Store, e *entity.Experiment) (*experimentReadView, error) {
-	views, err := annotateExperimentsForRead(s, []*entity.Experiment{e})
-	if err != nil {
-		return nil, err
-	}
-	if len(views) == 0 {
-		return nil, nil
-	}
-	return views[0], nil
+	return readmodel.AnnotateExperimentForRead(s, e)
 }
 
 func readExperimentForRead(s *store.Store, id string) (*experimentReadView, error) {
-	e, err := s.ReadExperiment(id)
-	if err != nil {
-		return nil, err
-	}
-	return annotateExperimentForRead(s, e)
+	return readmodel.ReadExperimentForRead(s, id)
 }
 
 func listScopedExperimentsForRead(s *store.Store, scope goalScope) ([]*experimentReadView, error) {
@@ -133,30 +58,11 @@ func listExperimentsForHypothesisForRead(s *store.Store, hypID string) ([]*exper
 }
 
 func classifyAllExperimentsForRead(s *store.Store) (map[string]experimentReadClass, error) {
-	exps, err := s.ListExperiments()
-	if err != nil {
-		return nil, err
-	}
-	return classifyExperimentsForRead(s, exps)
-}
-
-func classifyExperimentForRead(e *entity.Experiment, hypStatus map[string]string) experimentReadClass {
-	if e == nil {
-		return normalizeExperimentReadClass(experimentReadClass{})
-	}
-	return classifyHypothesisStatusForExperimentRead(hypStatus[e.Hypothesis])
+	return readmodel.ClassifyAllExperimentsForRead(s)
 }
 
 func classifyHypothesisStatusForExperimentRead(status string) experimentReadClass {
-	switch status {
-	case entity.StatusUnreviewed, entity.StatusSupported, entity.StatusRefuted, entity.StatusKilled:
-		return normalizeExperimentReadClass(experimentReadClass{
-			Classification:   experimentClassificationDead,
-			HypothesisStatus: status,
-		})
-	default:
-		return normalizeExperimentReadClass(experimentReadClass{})
-	}
+	return readmodel.ClassifyHypothesisStatusForExperimentRead(status)
 }
 
 func validateExperimentClassificationFilter(classification string) error {
@@ -169,7 +75,9 @@ func validateExperimentClassificationFilter(classification string) error {
 }
 
 func experimentClassificationSummary(classification, hypothesisStatus string) string {
-	classification = normalizeExperimentReadClass(experimentReadClass{Classification: classification}).Classification
+	if classification == "" {
+		classification = experimentClassificationLive
+	}
 	if classification != experimentClassificationDead {
 		return classification
 	}
@@ -187,34 +95,5 @@ func experimentClassificationMarker(classification string) string {
 }
 
 func findStaleExperimentsForRead(exps []*entity.Experiment, classByID map[string]experimentReadClass, events []store.Event, threshold time.Duration, now time.Time) []staleExperimentView {
-	stale := []staleExperimentView{}
-	for _, e := range exps {
-		switch e.Status {
-		case entity.ExpDesigned, entity.ExpImplemented, entity.ExpMeasured:
-		default:
-			continue
-		}
-		if e.IsBaseline {
-			continue
-		}
-		if !experimentReadClassForID(classByID, e.ID).loopActionable() {
-			continue
-		}
-		ts, kind := findLastEventForExperiment(events, e.ID)
-		if ts == nil {
-			continue
-		}
-		age := now.Sub(*ts)
-		if age < threshold {
-			continue
-		}
-		stale = append(stale, staleExperimentView{
-			ID:            e.ID,
-			Hypothesis:    e.Hypothesis,
-			Status:        e.Status,
-			LastEventKind: kind,
-			StaleMinutes:  age.Minutes(),
-		})
-	}
-	return stale
+	return readmodel.FindStaleExperimentsForRead(exps, classByID, events, threshold, now)
 }
